Correction du type d'identifiant dans DPIRepository pour utiliser Integer au lieu de Long

+++ back/src/main/java/esiag/back/repositories/DPIRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DPIRepository extends JpaRepository<DPI, Long> {
+public interface DPIRepository extends JpaRepository<DPI, Integer> {
 
 }
